requirements.txt
absl-py==2.0.0
altair==5.1.1
altgraph==0.17.4
appdirs==1.4.4
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.0
flatbuffers==23.5.26
fonttools==4.43.0
frozendict==2.3.9
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
multitasking==0.0.11
numpy==1.26.0
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
peewee==3.17.0
pefile==2023.2.7
Pillow==10.0.1
premailer==3.10.0
protobuf==4.24.3
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
pygame==2.3.0
Pygments==2.16.1
pyinstall==0.1.4
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
pynput==1.7.6
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.3
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sklearn==0.0.post11
smmap==5.0.1
soupsieve==2.5
streamlit==1.27.1
streamlit-option-menu==0.3.6
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tk==0.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
validators==0.22.0
watchdog==3.0.0
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
yagmail==0.15.293
yfinance==0.2.32
zipp==3.17.0
